pyproject.toml
[tool.poetry]
name = "depen_analyze"
version = "0.1.0"
description = ""
authors = ["even <evenh_1998@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
dnspython = "^2.2.1"
tldextract = "^3.4.0"
whois = "^0.9.21"
tqdm = "^4.64.1"
fire = "^0.5.0"
requests = "^2.28.1"
selenium = "^4.7.2"
matplotlib = "^3.6.2"
numpy = "^1.24.1"
jsbeautifier = "^1.14.7"
beautifulsoup4 = "^4.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
